adjust sample table generation and processing and defined outputs

+++ generate_sample_table.sh
@@ -13,7 +13,8 @@ HELP="""
 ASSUMES FASTQ FILES ARE NAMED AS <sample_name>_L00#_R{1,2}*.fastq(.gz)
 
 Flags:
-    -d  :  Path to directory containing sample fastq(.gz) files (Absolute paths preferred for consistency, but can use relative paths)
+    -d  :  Path to directory containing sample fastq(.gz) files (Absolute paths preferred for consistency, but can use relative paths) (incompatible with -b)
+    -b  :  Path to directory containing sample bam(.gz) files (incompatible with -d)
     -n  :  Name or file path for final sample table (with extension) (default: 'sample_table.csv') - will overwrite if file exists
     -e  :  Name or file path for an existing sample table - will append to the end of the provided table
 
@@ -21,24 +22,33 @@ Select one of '-n' (new sample table) or '-e' (existing sample table).
 If neither provided, a new sample table called 'sample_table.csv' will be created (or overwritten) by default. 
 """
 
-while getopts ":d:n:e:" option; do
+while getopts ":d:b:n:e:" option; do
     case "${option}" in
         d) database_dir=$OPTARG;;
+        b) bam_dir=$OPTARG;;
         n) name=$OPTARG;;
         e) existing=$OPTARG;;
     esac
 done
 
-if [ $database_dir = 0 ] ; then
-    echo "You must specify a data directory containing fastq(.gz) reads."
-    echo "$HELP"
-    exit 1
+if [ $database_dir = 0 ] && [ $bam_dir = 0 ]; then
+	echo "You must specify a data directory containing fastq(.gz) reads (-d) or bam(.gz) reads (-b)."
+	echo "$HELP"
+	exit 1
 fi
 
+if [ ! $database_dir = 0 ] && [ ! $bam_dir = 0 ]; then
+	echo "You can only select one directory of fastq(.gz) reads (-d) or bam(.gz) reads (-b)."
+	echo "$HELP"
+	exit 1
+
 echo -e "Adding samples\n"
-if [ $existing = 0 ] ; then
+if [ $existing = 0 ] && [ ! $database_dir = 0 ]; then
 	echo -e "Creating new sample table: ${name}\n"
- 	echo "sample,r1_path,r2_path" > ${name} && echo "sample,r1_path,r2_path"
+	echo "sample,r1_path,r2_path" > ${name} && echo "sample,r1_path,r2_path"
+elif [ $existing = 0 ] && [ ! $bam_dir = 0 ]; then
+	echo -e "Creating new sample table: ${name}\n"
+	echo "sample,bam_path" > ${name} && echo "sample,bam_path"
 else
 	filename=$(basename $existing)
 	if [ -f ""$existing"" ] ; then
@@ -77,6 +87,19 @@ if [ ! $database_dir = 0 ]; then
 	done
 	echo -e "\n"
 fi
+
+if [ ! $bam_dir = 0 ]; then
+		samples_dir=()
+		samples_fail=()
+		for file in $bam_dir/*.bam(.gz); do
+			sample=$(basename $file | cut -d_ -f 1 | cut -d. -f1)
+			if [ -f $file ]; then
+				echo ${sample},${file} >> ${name} && echo ${sample},${file}
+			else
+				samples_fail+=("${sample}")
+			fi
+		done
+		echo -e "\n"
 
 if [ ${#samples_fail[@]} -gt 0 ]; then 
 	echo -e "Samples that failed to be added to sample table:"
